refactor(captions): split into srt.ts + ass.ts + captions.ts
captions.ts (743 lines) split into:
  srt.ts (88 lines) — formatTime, convertToSRT, parseSrt, SrtEntry
  ass.ts (306 lines) — SubtitleStyle, STYLE_PRESETS, convertSrtToAss,
    convertSrtToAssGrouped, formatAssTime, colorToAss, EntryEmojiData
  captions.ts (370 lines) — renderCaptions orchestrator + helpers
    (resolveSrtContent, buildEmojiOverlays)

Each file has a single responsibility:
  srt.ts   — SRT format <-> structured data
  ass.ts   — SRT -> ASS conversion with styles + emoji data
  captions.ts — orchestration: transcription, font resolution, emoji overlays
